pyproject.toml
[project]
name = "picnic-topic-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.0.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "optuna>=4.5.0",
    "optuna-integration>=4.5.0",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
]
